requirements.txt
openai>=1.30.0
huggingface_hub
python-dotenv>=1.0.1